requirements.txt
Django==3.2.10
asgiref==3.4.1
backports.zoneinfo==0.2.1
django-rest-auth==0.9.5
djangorestframework==3.13.1
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
Pillow==9.0.0
